requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
sqlalchemy==2.0.46
pymysql==1.1.2
pydantic_settings==2.12.0
PyJWT==2.10.1
passlib==1.7.4
python-multipart==0.0.22
bcrypt==4.0.1
alembic==1.18.3
pycountry==24.6.1
aiomysql==0.3.2
pytest==9.0.2
cryptography==46.0.4
httpx==0.28.1




